Сделали отрисовку чарта классом
