--- src/filters.js
@@ -1,27 +1,0 @@
-export const filtersList = [
-  {
-    'anchor': `all`,
-    'state': `active`,
-    'label': `All movies`
-  },
-  {
-    'anchor': `whatchlist`,
-    'label': `Whatchlist`,
-    'count': 13
-  },
-  {
-    'anchor': `history`,
-    'label': `History`,
-    'count': 4
-  },
-  // {
-  //   'anchor': `favorites`,
-  //   'label': `Favorites`,
-  //   'count': 8
-  // },
-  // {
-  //   'anchor': `stats`,
-  //   'state': `additional`,
-  //   'label': `Stats`
-  // }
-];
